Moving getPrompts from utils folder
