+++ src/compile.js
@@ -2,6 +2,7 @@ const fs = require('fs');
 const path = require('path');
 const { node } = require('compile-run');
 const { Utils, CompilationError } = require('./utils/index.js');
+const  { getPrompts } = require("./utils");
 
 module.exports = {
   validate: () => true,
@@ -12,7 +13,7 @@ module.exports = {
 
     const content = fs.readFileSync(entryPath, "utf8");
 
-    const promptsValues = Utils.getPrompts(content);
+    const promptsValues = getPrompts(content);
 
     const inputs = (promptsValues.length === 0) ? [] : await socket.ask(promptsValues);
